Streamline design of headers and component utils
- Remove page title from navbar
- Add unified header to every view class
- Re-enable live reload of frontend resources
