+++ src/main/java/de/sustineo/simdesk/views/SessionDetailsView.java
@@ -8,36 +8,31 @@ import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.grid.ColumnTextAlign;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.grid.GridVariant;
+import com.vaadin.flow.component.grid.dataview.GridListDataView;
 import com.vaadin.flow.component.html.Anchor;
 import com.vaadin.flow.component.html.H3;
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import com.vaadin.flow.component.tabs.Tab;
-import com.vaadin.flow.component.tabs.TabSheet;
-import com.vaadin.flow.component.tabs.TabSheetVariant;
-import com.vaadin.flow.data.renderer.LitRenderer;
+import com.vaadin.flow.data.selection.SingleSelect;
 import com.vaadin.flow.router.*;
 import com.vaadin.flow.server.StreamResource;
 import com.vaadin.flow.server.auth.AnonymousAllowed;
 import de.sustineo.simdesk.configuration.ProfileManager;
-import de.sustineo.simdesk.configuration.VaadinConfiguration;
-import de.sustineo.simdesk.entities.Car;
-import de.sustineo.simdesk.entities.Lap;
-import de.sustineo.simdesk.entities.Penalty;
 import de.sustineo.simdesk.entities.Session;
+import de.sustineo.simdesk.entities.SessionType;
 import de.sustineo.simdesk.entities.auth.Role;
+import de.sustineo.simdesk.entities.comparator.SessionRankingLapTimeComparator;
+import de.sustineo.simdesk.entities.ranking.SessionRanking;
 import de.sustineo.simdesk.layouts.MainLayout;
 import de.sustineo.simdesk.services.auth.SecurityService;
-import de.sustineo.simdesk.services.leaderboard.LapService;
-import de.sustineo.simdesk.services.leaderboard.PenaltyService;
 import de.sustineo.simdesk.services.leaderboard.RankingService;
 import de.sustineo.simdesk.services.leaderboard.SessionService;
 import de.sustineo.simdesk.utils.FormatUtils;
-import de.sustineo.simdesk.views.generators.InvalidLapPartNameGenerator;
-import de.sustineo.simdesk.views.renderers.LapRenderer;
-import de.sustineo.simdesk.views.renderers.SessionDetailsRenderer;
+import de.sustineo.simdesk.views.generators.SessionRankingDNFNameGenerator;
+import de.sustineo.simdesk.views.generators.SessionRankingPodiumPartNameGenerator;
+import de.sustineo.simdesk.views.renderers.SessionRankingRenderer;
 import lombok.extern.java.Log;
 import org.springframework.context.annotation.Profile;
 
@@ -45,34 +40,26 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Log
 @Profile(ProfileManager.PROFILE_LEADERBOARD)
-@Route(value = "/leaderboard/sessions/:fileChecksum/details/:carId", layout = MainLayout.class)
-@PageTitle(VaadinConfiguration.APPLICATION_NAME_PREFIX + "Leaderboard - Session")
+@Route(value = "/leaderboard/sessions/:fileChecksum", layout = MainLayout.class)
+@PageTitle("Leaderboard - Session Details")
 @AnonymousAllowed
 public class SessionDetailsView extends BaseView implements BeforeEnterObserver {
-    private final SessionService sessionService;
-    private final LapService lapService;
-    private final PenaltyService penaltyService;
     private final RankingService rankingService;
+    private final SessionService sessionService;
     private final SecurityService securityService;
 
-    private List<Lap> laps = new ArrayList<>();
-    private List<Penalty> penalties = new ArrayList<>();
+    private GridListDataView<SessionRanking> dataView;
 
-    public SessionDetailsView(SessionService sessionService,
-                              LapService lapService,
-                              PenaltyService penaltyService,
-                              RankingService rankingService,
+    public SessionDetailsView(RankingService rankingService,
+                              SessionService sessionService,
                               SecurityService securityService) {
-        this.sessionService = sessionService;
-        this.lapService = lapService;
-        this.penaltyService = penaltyService;
         this.rankingService = rankingService;
+        this.sessionService = sessionService;
         this.securityService = securityService;
 
         setSizeFull();
@@ -82,37 +69,22 @@ public class SessionDetailsView extends BaseView implements BeforeEnterObserver 
 
     @Override
     public void beforeEnter(BeforeEnterEvent beforeEnterEvent) {
-        RouteParameters routeParameters = beforeEnterEvent.getRouteParameters();
+        final RouteParameters routeParameters = beforeEnterEvent.getRouteParameters();
 
         String fileChecksum = routeParameters.get(ROUTE_PARAMETER_FILE_CHECKSUM).orElseThrow();
-        int carId = Integer.parseInt(routeParameters.get(ROUTE_PARAMETER_CAR_ID).orElseThrow());
 
-        Session session = sessionService.getSession(fileChecksum);
-        if (session == null) {
-            throw new NotFoundException("Session with file checksum " + fileChecksum + " does not exist.");
+        try {
+            Session session = sessionService.getSession(fileChecksum);
+            if (session == null) {
+                throw new IllegalArgumentException("Session with file checksum " + fileChecksum + " does not exist.");
+            }
+
+            add(createViewHeader());
+            add(createSessionInformation(session));
+            addAndExpand(createLeaderboardGrid(session));
+        } catch (IllegalArgumentException e) {
+            beforeEnterEvent.rerouteToError(NotFoundException.class);
         }
-
-        List<String> playerIds = rankingService.getPlayerIdsBySessionAndCarId(session.getId(), carId);
-        if (playerIds == null || playerIds.isEmpty()) {
-            throw new NotFoundException("No car found in session with file checksum " + fileChecksum + " and car id " + carId);
-        }
-
-        laps = lapService.getLapsBySessionAndDrivers(session.getId(), playerIds);
-        penalties = penaltyService.findBySessionAndCarId(session.getId(), carId).stream()
-                .filter(Penalty::isValid)
-                .collect(Collectors.toList());
-
-        add(createSessionInformation(session));
-
-        TabSheet tabSheet = new TabSheet();
-        tabSheet.setSizeFull();
-        tabSheet.addThemeVariants(TabSheetVariant.LUMO_TABS_CENTERED);
-        tabSheet.add(createTab("Laps", laps.size()), createLapsGrid());
-        tabSheet.add(createTab("Penalties", penalties.size()), createPenaltyGrid());
-        if (securityService.hasAnyRole(Role.ADMIN)) {
-            tabSheet.add(createTab("Statistics"), createStatisticsLayout(fileChecksum, carId));
-        }
-        addAndExpand(tabSheet);
     }
 
     private Component createSessionInformation(Session session) {
@@ -124,174 +96,156 @@ public class SessionDetailsView extends BaseView implements BeforeEnterObserver 
         H3 heading = new H3();
         heading.setText(String.format("%s - %s - %s", session.getSessionType().getDescription(), session.getTrackName(), session.getServerName()));
 
-        Icon weatherIcon = ComponentUtils.createWeatherIcon(session);
+        Icon weatherIcon = getWeatherIcon(session);
 
         Span sessionDatetimeBadge = new Span();
         sessionDatetimeBadge.setText(FormatUtils.formatDatetime(session.getSessionDatetime()));
         sessionDatetimeBadge.getElement().getThemeList().add("badge contrast");
 
         layout.add(weatherIcon, heading, sessionDatetimeBadge);
+
+        if (securityService.hasAnyRole(Role.ADMIN)) {
+            StreamResource csvResource = new StreamResource(
+                    String.format("session_table_%s.csv", session.getFileChecksum()),
+                    () -> {
+                        String csv = this.exportCSV();
+                        return new ByteArrayInputStream(csv != null ? csv.getBytes(StandardCharsets.UTF_8) : new byte[0]);
+                    }
+            );
+            Anchor downloadSessionAnchor = createDownloadAnchor(csvResource, "Table (CSV)");
+
+            StreamResource fileContentResource = new StreamResource(
+                    String.format("session_file_%s.json", session.getFileChecksum()),
+                    () -> {
+                        String fileContent = session.getFileContent();
+                        return new ByteArrayInputStream(fileContent != null ? fileContent.getBytes(StandardCharsets.UTF_8) : new byte[0]);
+                    }
+            );
+            Anchor downloadSessionFileAnchor = createDownloadAnchor(fileContentResource, "File (JSON)");
+
+            layout.add(downloadSessionAnchor, downloadSessionFileAnchor);
+        }
+
         return layout;
     }
 
-    private Component createLapsGrid() {
-        Grid<Lap> grid = new Grid<>(Lap.class, false);
-        grid.addColumn(LitRenderer.of("${index + 1}"))
+    private Component createLeaderboardGrid(Session session) {
+        List<SessionRanking> sessionRankings = rankingService.getSessionRankings(session);
+        List<SessionRanking> validSessionRankings = sessionRankings.stream()
+                .filter(SessionRanking::isValid)
+                .toList();
+
+        if (validSessionRankings.isEmpty()) {
+            VerticalLayout layout = new VerticalLayout();
+
+            layout.setPadding(true);
+            layout.setWidthFull();
+            layout.setAlignItems(Alignment.CENTER);
+
+            H3 message = new H3("No valid laps in this session!");
+            message.getStyle()
+                    .setColor("var(--lumo-error-text-color)");
+
+            layout.add(message);
+            return layout;
+        }
+
+        SessionRanking bestTotalTimeSessionRanking = validSessionRankings.stream().findFirst().orElse(new SessionRanking());
+        SessionRanking bestLapTimeSessionRanking = validSessionRankings.stream()
+                .min(new SessionRankingLapTimeComparator())
+                .orElse(new SessionRanking());
+
+        Grid<SessionRanking> grid = new Grid<>(SessionRanking.class, false);
+        grid.addColumn(SessionRanking::getRanking)
                 .setHeader("#")
-                .setWidth(ComponentUtils.GRID_RANKING_WIDTH)
+                .setWidth(GRID_RANKING_WIDTH)
                 .setFlexGrow(0)
                 .setSortable(true)
                 .setTextAlign(ColumnTextAlign.CENTER)
-                .setFrozen(true);
-        grid.addColumn(lap -> Car.getCarGroupById(lap.getCarModelId()))
+                .setPartNameGenerator(new SessionRankingPodiumPartNameGenerator());
+        grid.addColumn(SessionRankingRenderer.createRaceNumberRenderer())
+                .setHeader("Race Number")
+                .setAutoWidth(true)
+                .setFlexGrow(0)
+                .setSortable(true)
+                .setComparator(SessionRanking::getRaceNumber);
+        grid.addColumn(SessionRanking::getCarGroup)
                 .setHeader("Car Group")
                 .setAutoWidth(true)
                 .setFlexGrow(0)
                 .setSortable(true);
-        grid.addColumn(lap -> Car.getCarNameById(lap.getCarModelId()))
+        grid.addColumn(SessionRanking::getCarModelName)
                 .setHeader("Car Model")
                 .setAutoWidth(true)
                 .setFlexGrow(0)
-                .setSortable(true)
-                .setComparator(Lap::getLapTimeMillis);
-        grid.addColumn(lap -> lap.getDriver().getFullName())
-                .setHeader("Driver")
                 .setSortable(true);
-        grid.addColumn(LapRenderer.createLapTimeRenderer())
-                .setHeader("Lap Time")
+        grid.addColumn(SessionRankingRenderer.createDriversRenderer())
+                .setHeader("Drivers")
+                .setSortable(true);
+        grid.addColumn(SessionRanking::getLapCount)
+                .setHeader("Laps")
                 .setAutoWidth(true)
                 .setFlexGrow(0)
-                .setPartNameGenerator(ranking -> "font-weight-bold")
-                .setSortable(true)
-                .setComparator(Lap::getLapTimeMillis);
-        grid.addColumn(LapRenderer.createSplit1Renderer())
-                .setHeader("Split 1")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setSortable(true)
-                .setComparator(Lap::getSplit1Millis);
-        grid.addColumn(LapRenderer.createSplit2Renderer())
-                .setHeader("Split 2")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setSortable(true)
-                .setComparator(Lap::getSplit2Millis);
-        grid.addColumn(LapRenderer.createSplit3Renderer())
-                .setHeader("Split 3")
+                .setSortable(true);
+        if (SessionType.R.equals(bestTotalTimeSessionRanking.getSession().getSessionType())) {
+            grid.addColumn(SessionRankingRenderer.createTotalTimeRenderer(bestTotalTimeSessionRanking))
+                    .setHeader("Total Time")
+                    .setAutoWidth(true)
+                    .setFlexGrow(0)
+                    .setSortable(true)
+                    .setComparator(SessionRanking::getTotalTimeMillis);
+        }
+        grid.addColumn(SessionRankingRenderer.createLapTimeRenderer(bestLapTimeSessionRanking))
+                .setHeader("Fastest Lap")
                 .setAutoWidth(true)
                 .setFlexGrow(0)
                 .setSortable(true)
-                .setComparator(Lap::getSplit3Millis);
+                .setComparator(SessionRanking::getBestLapTimeMillis);
 
-        grid.setItems(laps);
+        dataView = grid.setItems(validSessionRankings);
         grid.setHeightFull();
         grid.setMultiSort(true, true);
         grid.setColumnReorderingAllowed(true);
         grid.addThemeVariants(GridVariant.LUMO_NO_BORDER);
         grid.setSelectionMode(Grid.SelectionMode.NONE);
-        grid.setPartNameGenerator(new InvalidLapPartNameGenerator());
+        grid.setPartNameGenerator(new SessionRankingDNFNameGenerator(bestTotalTimeSessionRanking));
+        grid.setSelectionMode(Grid.SelectionMode.SINGLE);
+
+        SingleSelect<Grid<SessionRanking>, SessionRanking> singleSelect = grid.asSingleSelect();
+        singleSelect.addValueChangeListener(e -> {
+            SessionRanking selectedSessionRanking = e.getValue();
+
+            if (selectedSessionRanking != null) {
+                getUI().ifPresent(ui -> ui.navigate(SessionCarDetailsView.class,
+                        new RouteParameters(
+                                new RouteParam(ROUTE_PARAMETER_FILE_CHECKSUM, session.getFileChecksum()),
+                                new RouteParam(ROUTE_PARAMETER_CAR_ID, selectedSessionRanking.getCarId().toString())
+                        )
+                ));
+            }
+        });
 
         return grid;
     }
 
-    private Component createPenaltyGrid() {
-        Grid<Penalty> grid = new Grid<>(Penalty.class, false);
-        grid.addColumn(LitRenderer.of("${index + 1}"))
-                .setHeader("#")
-                .setWidth(ComponentUtils.GRID_RANKING_WIDTH)
-                .setFlexGrow(0)
-                .setSortable(true)
-                .setTextAlign(ColumnTextAlign.CENTER)
-                .setFrozen(true);
-        grid.addColumn(Penalty::getPenaltyAbbreviation)
-                .setHeader("Penalty")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setSortable(true);
-        grid.addColumn(Penalty::getReasonDescription)
-                .setHeader("Reason")
-                .setSortable(true);
-        grid.addColumn(Penalty::getViolationLapCorrected)
-                .setHeader("Violation in lap")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setSortable(true);
-        grid.addColumn(Penalty::getClearedLapCorrected)
-                .setHeader("Cleared in lap")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setSortable(true);
-        grid.addColumn(SessionDetailsRenderer.createPenaltyServedRenderer())
-                .setHeader("Cleared")
-                .setAutoWidth(true)
-                .setFlexGrow(0)
-                .setTextAlign(ColumnTextAlign.END)
-                .setSortable(true);
-
-        grid.setItems(penalties);
-        grid.setHeightFull();
-        grid.setMultiSort(true, true);
-        grid.setColumnReorderingAllowed(true);
-        grid.addThemeVariants(GridVariant.LUMO_NO_BORDER);
-        grid.setSelectionMode(Grid.SelectionMode.NONE);
-
-        return grid;
-    }
-
-    private Component createStatisticsLayout(String fileChecksum, int carId) {
-        VerticalLayout layout = new VerticalLayout();
-
-        StreamResource csvResource = new StreamResource(
-                String.format("session_laps_%s_%s.csv", fileChecksum, carId),
-                () -> {
-                    String csv = this.exportLapsCSV();
-                    return new ByteArrayInputStream(csv != null ? csv.getBytes(StandardCharsets.UTF_8) : new byte[0]);
-                }
-        );
-
-        Anchor downloadLapsAnchor = ComponentUtils.createDownloadAnchor(csvResource, "Download laps");
-
-        layout.add(downloadLapsAnchor);
-        return layout;
-    }
-
-    private String exportLapsCSV() {
-        if (laps == null) {
+    private String exportCSV() {
+        if (dataView == null) {
             return null;
         }
 
+        Stream<SessionRanking> sessionRows = dataView.getItems();
+
         try (StringWriter writer = new StringWriter()) {
-            StatefulBeanToCsv<Lap> sbc = new StatefulBeanToCsvBuilder<Lap>(writer)
+            StatefulBeanToCsv<SessionRanking> sbc = new StatefulBeanToCsvBuilder<SessionRanking>(writer)
                     .withSeparator(';')
                     .build();
-            sbc.write(laps);
+
+            sbc.write(sessionRows);
 
             return writer.toString();
         } catch (CsvDataTypeMismatchException | CsvRequiredFieldEmptyException | IOException e) {
             log.severe("An error occurred during creation of CSV resource: " + e.getMessage());
             return null;
         }
-    }
-
-    private Tab createTab(String label) {
-        return createTab(label, null);
-    }
-
-    private Tab createTab(String label, Integer number) {
-        HorizontalLayout horizontalLayout = new HorizontalLayout();
-        horizontalLayout.setAlignItems(Alignment.CENTER);
-        horizontalLayout.getStyle()
-                .set("gap", "var(--lumo-space-s)");
-
-        Span labelSpan = new Span(label);
-        horizontalLayout.add(labelSpan);
-
-        if (number != null) {
-            Span numberSpan = new Span(String.valueOf(number));
-            numberSpan.getElement().getThemeList().add("badge contrast");
-            horizontalLayout.add(numberSpan);
-        }
-
-        return new Tab(horizontalLayout);
     }
 }
